Code moving + better icon
