+++ entities2/CMakeLists.txt
@@ -55,11 +55,13 @@ set(
 set(
     ENTITIES2_SOURCE
     ${CMAKE_SOURCE_DIR}/src/ai.cpp
+    ${CMAKE_SOURCE_DIR}/src/difficulty_picker.cpp
     ${CMAKE_SOURCE_DIR}/src/discord_rpc.cpp
     ${CMAKE_SOURCE_DIR}/src/energy.cpp
     ${CMAKE_SOURCE_DIR}/src/entity.cpp
     ${CMAKE_SOURCE_DIR}/src/exit_msg.cpp
     ${CMAKE_SOURCE_DIR}/src/gameplay_info.cpp
+    ${CMAKE_SOURCE_DIR}/src/gameplay_loop.cpp
     ${CMAKE_SOURCE_DIR}/src/gen_moves.cpp
     ${CMAKE_SOURCE_DIR}/src/main.cpp
     ${CMAKE_SOURCE_DIR}/src/pick_move.cpp
