+++ entities2/src/gen_moves.cpp
@@ -145,6 +145,7 @@ static void PrintMoves(uint32_t* moves, uint32_t* move_types, double* energies)
                 break;
         }
     }
+    fmt::print("\n");
 }
 
 /**
