+++ entities2/src/headers/difficulty_picker.hpp
@@ -1,0 +1,5 @@
+// entities2 © 2024 by norbcodes is licensed under CC BY-NC 4.0
+
+#pragma once
+
+void DifficultyPicker();
