+++ entities2/src/headers/gameplay_loop.hpp
@@ -1,0 +1,7 @@
+// entities2 © 2024 by norbcodes is licensed under CC BY-NC 4.0
+
+#pragma once
+
+#include <cstdint>
+
+void Game(uint32_t mode, uint32_t& picker_flag);
