+++ entities2/src/windows/e2_resources.txt
@@ -2,7 +2,5 @@
 #include <afxres.h>
 
 #define IDI_ICON1 1
-#define IDI_ICON2 2
 
-IDI_ICON1 ICON "${CMAKE_SOURCE_DIR}/assets/icon/entities2_icon_128x128_ico.ico"
-IDI_ICON2 ICON "${CMAKE_SOURCE_DIR}/assets/icon/entities2_icon_256x256_ico.ico"
+IDI_ICON1 ICON "${CMAKE_SOURCE_DIR}/assets/icon/entities2_icon.ico"
